fix: prevent invalid unsigned difference arithmetic
`len` and `pref_len` are both unsigned integers here. Comparing their
difference `len - pref_len > 0` is a bit weird: we'll never be _less_
than zero, so really this check is either looking for:
* len - pref_len == 0
* overflowing when pref_len > len and giving us the incorrect result

This _is_ a fix that's included in later versions of protobuf-c, but
`iot-device-embedded-sdk-c` has been abandoned by Google and is not
seeing any new development.

Popping this commit in to silence some CodeQL alerts.