docs: document fork/tenera branch
This empty commit simply documents the `fork/tenera` branch.

This branch will be a dumping ground for any Tenera-specific changes we
make to `iot-device-embedded-sdk-c` that we don't intend to propose as a
PR to ClearBlade.

We're going to keep this branch alive in perpetuity: our internal PRs
will target (and get merged into) this branch. We'll then periodically
merge this branch into our `master` branch to provide longer-term merge
commits that we can target in git submodules.

This should allow us to:
a) cleanly isolate internal changes to our own branch to simplify
   creating separate branches to contribute back to ClearBlade
b) preserve a long-term history of how this fork evolved so that git
   submodule references in other projects remain valid